+++ workflow-tracing-papa/src/main/java/com/squareup/workflow1/tracing/WorkflowTrace.kt
@@ -1,0 +1,44 @@
+package com.squareup.workflow1.tracing
+
+import androidx.tracing.Trace
+import androidx.tracing.trace
+
+/**
+ * Production implementation of [TraceInterface] that uses androidx.tracing.Trace.
+ *
+ * @param isTraceable Whether tracing is enabled. Clients should configure this directly.
+ *   Defaults to false for backwards compatibility.
+ */
+class WorkflowTrace(
+  override val isTraceable: Boolean = false
+) : TraceInterface {
+
+  override val isCurrentlyTracing: Boolean
+    get() = Trace.isEnabled()
+
+  override fun beginSection(label: String) {
+    Trace.beginSection(label)
+  }
+
+  override fun endSection() {
+    Trace.endSection()
+  }
+
+  override fun beginAsyncSection(
+    name: String,
+    cookie: Int
+  ) {
+    Trace.beginAsyncSection(name, cookie)
+  }
+
+  override fun endAsyncSection(
+    name: String,
+    cookie: Int
+  ) {
+    Trace.endAsyncSection(name, cookie)
+  }
+
+  override fun logSection(info: String) {
+    trace(info) {}
+  }
+}
