+++ workflow-tracing-papa/src/main/java/com/squareup/workflow1/tracing/papa/PapaSafeTrace.kt
@@ -1,45 +1,36 @@
+@file:Suppress("DEPRECATION")
+
 package com.squareup.workflow1.tracing.papa
 
-import androidx.tracing.Trace
-import androidx.tracing.trace
 import com.squareup.workflow1.tracing.SafeTraceInterface
+import com.squareup.workflow1.tracing.WorkflowTrace
 
-/**
- * Production implementation of [SafeTraceInterface] that uses androidx.tracing.Trace.
- *
- * @param isTraceable Whether tracing is enabled. Clients should configure this directly.
- *   Defaults to false for backwards compatibility.
- */
+@Deprecated(
+  message = "Renamed to WorkflowTrace and moved to com.squareup.workflow1.tracing package",
+  replaceWith = ReplaceWith(
+    expression = "WorkflowTrace",
+    imports = arrayOf("com.squareup.workflow1.tracing.WorkflowTrace")
+  )
+)
 class PapaSafeTrace(
-  override val isTraceable: Boolean = false
+  isTraceable: Boolean = false
 ) : SafeTraceInterface {
+  private val delegate = WorkflowTrace(isTraceable)
 
-  override val isCurrentlyTracing: Boolean
-    get() = Trace.isEnabled()
+  override val isTraceable: Boolean get() = delegate.isTraceable
+  override val isCurrentlyTracing: Boolean get() = delegate.isCurrentlyTracing
 
-  override fun beginSection(label: String) {
-    Trace.beginSection(label)
-  }
-
-  override fun endSection() {
-    Trace.endSection()
-  }
-
+  override fun beginSection(label: String) = delegate.beginSection(label)
+  override fun endSection() = delegate.endSection()
   override fun beginAsyncSection(
     name: String,
     cookie: Int
-  ) {
-    Trace.beginAsyncSection(name, cookie)
-  }
+  ) = delegate.beginAsyncSection(name, cookie)
 
   override fun endAsyncSection(
     name: String,
     cookie: Int
-  ) {
-    Trace.endAsyncSection(name, cookie)
-  }
+  ) = delegate.endAsyncSection(name, cookie)
 
-  override fun logSection(info: String) {
-    trace(info) {}
-  }
+  override fun logSection(info: String) = delegate.logSection(info)
 }
