+++ workflow-tracing-papa/src/test/java/com/squareup/workflow1/tracing/papa/FakeSafeTrace.kt
@@ -1,53 +1,41 @@
+@file:Suppress("DEPRECATION")
+
 package com.squareup.workflow1.tracing.papa
 
+import com.squareup.workflow1.tracing.FakeTrace
 import com.squareup.workflow1.tracing.SafeTraceInterface
 
-/**
- * Fake implementation of [SafeTraceInterface] for testing purposes.
- * Records all trace calls for verification in tests.
- */
-class FakeSafeTrace(
-  override val isTraceable: Boolean = true,
-  override val isCurrentlyTracing: Boolean = true
-) : SafeTraceInterface {
-
-  data class TraceCall(
-    val type: String,
-    val label: String? = null,
-    val name: String? = null,
-    val cookie: Int? = null
+@Deprecated(
+  message = "Renamed to FakeTrace and moved to com.squareup.workflow1.tracing package",
+  replaceWith = ReplaceWith(
+    expression = "FakeTrace",
+    imports = arrayOf("com.squareup.workflow1.tracing.FakeTrace")
   )
+)
+class FakeSafeTrace(
+  isTraceable: Boolean = true,
+  isCurrentlyTracing: Boolean = true
+) : SafeTraceInterface {
+  private val delegate = FakeTrace(isTraceable, isCurrentlyTracing)
 
-  private val _traceCalls = mutableListOf<TraceCall>()
-  val traceCalls: List<TraceCall> get() = _traceCalls.toList()
+  // These aren't part of TraceInterface
+  val traceCalls: List<FakeTrace.TraceCall> get() = delegate.traceCalls
+  fun clearTraceCalls() = delegate.clearTraceCalls()
 
-  fun clearTraceCalls() {
-    _traceCalls.clear()
-  }
+  override val isTraceable: Boolean get() = delegate.isTraceable
+  override val isCurrentlyTracing: Boolean get() = delegate.isCurrentlyTracing
 
-  override fun beginSection(label: String) {
-    _traceCalls.add(TraceCall("beginSection", label = label))
-  }
-
-  override fun endSection() {
-    _traceCalls.add(TraceCall("endSection"))
-  }
-
+  override fun beginSection(label: String) = delegate.beginSection(label)
+  override fun endSection() = delegate.endSection()
   override fun beginAsyncSection(
     name: String,
     cookie: Int
-  ) {
-    _traceCalls.add(TraceCall("beginAsyncSection", name = name, cookie = cookie))
-  }
+  ) = delegate.beginAsyncSection(name, cookie)
 
   override fun endAsyncSection(
     name: String,
     cookie: Int
-  ) {
-    _traceCalls.add(TraceCall("endAsyncSection", name = name, cookie = cookie))
-  }
+  ) = delegate.endAsyncSection(name, cookie)
 
-  override fun logSection(info: String) {
-    _traceCalls.add(TraceCall("logSection", label = info))
-  }
+  override fun logSection(info: String) = delegate.logSection(info)
 }
