+++ workflow-tracing-papa/src/test/java/com/squareup/workflow1/tracing/papa/WorkflowPapaTracerTest.kt
@@ -11,10 +11,13 @@ import com.squareup.workflow1.WorkflowInterceptor.RuntimeSettled
 import com.squareup.workflow1.WorkflowInterceptor.WorkflowSession
 import com.squareup.workflow1.identifier
 import com.squareup.workflow1.tracing.ConfigSnapshot
+import com.squareup.workflow1.tracing.FakeTrace
 import com.squareup.workflow1.tracing.RenderCause
 import com.squareup.workflow1.tracing.RuntimeTraceContext
 import com.squareup.workflow1.tracing.RuntimeUpdateLogLine
 import com.squareup.workflow1.tracing.WorkflowSessionInfo
+import com.squareup.workflow1.tracing.WorkflowTrace
+import com.squareup.workflow1.tracing.WorkflowTracer
 import kotlinx.coroutines.test.TestScope
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
@@ -25,8 +28,8 @@ import kotlin.test.assertTrue
 
 internal class WorkflowPapaTracerTest {
 
-  private val fakeTrace = FakeSafeTrace()
-  private val papaTracer = WorkflowPapaTracer(fakeTrace)
+  private val fakeTrace = FakeTrace()
+  private val tracer = WorkflowTracer(fakeTrace)
 
   @Test
   fun `onWorkflowSessionStarted creates async section for root workflow`() {
@@ -36,12 +39,12 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     // Add session info to the context as would normally be done by WorkflowRuntimeMonitor
     testContext.workflowSessionInfo[rootSession.sessionId] = WorkflowSessionInfo(rootSession)
 
-    papaTracer.onWorkflowSessionStarted(testScope, rootSession)
+    tracer.onWorkflowSessionStarted(testScope, rootSession)
 
     val asyncSectionCalls = fakeTrace.traceCalls.filter { it.type == "beginAsyncSection" }
     assertEquals(1, asyncSectionCalls.size)
@@ -56,12 +59,12 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     // Add session info to the context as would normally be done by WorkflowRuntimeMonitor
     testContext.workflowSessionInfo[rootSession.sessionId] = WorkflowSessionInfo(rootSession)
 
-    val result = papaTracer.onInitialState(
+    val result = tracer.onInitialState(
       props = "testProps",
       snapshot = null,
       workflowScope = testScope,
@@ -83,12 +86,12 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     val expectedSnapshot = TreeSnapshot.forRootOnly(null)
     val renderingAndSnapshot = RenderingAndSnapshot("rendering", expectedSnapshot)
 
-    val result = papaTracer.onRenderAndSnapshot(
+    val result = tracer.onRenderAndSnapshot(
       renderProps = "props",
       proceed = { renderingAndSnapshot },
       session = rootSession
@@ -101,23 +104,53 @@ internal class WorkflowPapaTracerTest {
 
   @Test
   fun `tracer can be instantiated`() {
-    assertNotNull(papaTracer)
+    assertNotNull(tracer)
   }
 
   @Test
-  fun `PapaSafeTrace can be configured with isTraceable`() {
-    val traceableTrace = PapaSafeTrace(isTraceable = true)
+  fun `WorkflowTrace can be configured with isTraceable`() {
+    val traceableTrace = WorkflowTrace(isTraceable = true)
     assertEquals(true, traceableTrace.isTraceable)
 
-    val nonTraceableTrace = PapaSafeTrace(isTraceable = false)
+    val nonTraceableTrace = WorkflowTrace(isTraceable = false)
     assertEquals(false, nonTraceableTrace.isTraceable)
   }
 
   @Test
-  fun `WorkflowPapaTracer can be configured with custom SafeTrace`() {
+  fun `WorkflowTracer can be configured with custom TraceInterface`() {
+    val customTrace = FakeTrace(isTraceable = true)
+    val tracer = WorkflowTracer(trace = customTrace)
+    assertNotNull(tracer)
+  }
+
+  @Suppress("DEPRECATION")
+  @Test
+  fun `deprecated WorkflowPapaTracer can be configured with custom SafeTraceInterface`() {
     val customTrace = FakeSafeTrace(isTraceable = true)
     val tracer = WorkflowPapaTracer(safeTrace = customTrace)
-    assertNotNull(tracer)
+
+    tracer.attachRuntimeContext(TestRuntimeTraceContext())
+    tracer.onSnapshotStateWithChildren(
+      proceed = { TreeSnapshot.forRootOnly(null) },
+      session = TestWorkflow().createMockSession()
+    )
+
+    assertTrue(customTrace.traceCalls.any { it.type == "beginSection" })
+  }
+
+  @Suppress("DEPRECATION")
+  @Test
+  fun `deprecated WorkflowPapaTracer can be configured with custom TraceInterface`() {
+    val customTrace = FakeTrace(isTraceable = true)
+    val tracer = WorkflowPapaTracer(safeTrace = customTrace)
+
+    tracer.attachRuntimeContext(TestRuntimeTraceContext())
+    tracer.onSnapshotStateWithChildren(
+      proceed = { TreeSnapshot.forRootOnly(null) },
+      session = TestWorkflow().createMockSession()
+    )
+
+    assertTrue(customTrace.traceCalls.any { it.type == "beginSection" })
   }
 
   @Test
@@ -127,12 +160,12 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     // Add session info to the context as would normally be done by WorkflowRuntimeMonitor
     testContext.workflowSessionInfo[mockSession.sessionId] = WorkflowSessionInfo(mockSession)
 
-    val result = papaTracer.onPropsChanged(
+    val result = tracer.onPropsChanged(
       old = "old",
       new = "new",
       state = "current",
@@ -150,12 +183,12 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     val expectedSnapshot = TreeSnapshot.forRootOnly(null)
     val renderingAndSnapshot = RenderingAndSnapshot("rendering", expectedSnapshot)
 
-    val result = papaTracer.onRenderAndSnapshot(
+    val result = tracer.onRenderAndSnapshot(
       renderProps = "props",
       proceed = { renderingAndSnapshot },
       session = mockSession
@@ -171,11 +204,11 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     val treeSnapshot = TreeSnapshot.forRootOnly(null)
 
-    val result = papaTracer.onSnapshotStateWithChildren(
+    val result = tracer.onSnapshotStateWithChildren(
       proceed = { treeSnapshot },
       session = mockSession
     )
@@ -186,15 +219,15 @@ internal class WorkflowPapaTracerTest {
   @Test
   fun `onRuntimeUpdateEnhanced handles different runtime updates`() {
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     val configSnapshot = ConfigSnapshot(TestRuntimeConfig())
 
     // Should not throw for RenderPassSkipped
-    papaTracer.onRuntimeUpdateEnhanced(RenderPassSkipped, false, configSnapshot)
+    tracer.onRuntimeUpdateEnhanced(RenderPassSkipped, false, configSnapshot)
 
     // Should not throw for RuntimeLoopSettled
-    papaTracer.onRuntimeUpdateEnhanced(RuntimeSettled, true, configSnapshot)
+    tracer.onRuntimeUpdateEnhanced(RuntimeSettled, true, configSnapshot)
   }
 
   @Test
@@ -204,14 +237,14 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     // Add session info to the context as would normally be done by WorkflowRuntimeMonitor
     testContext.workflowSessionInfo[mockSession.sessionId] = WorkflowSessionInfo(mockSession)
 
     // Should not throw
-    papaTracer.onWorkflowSessionStarted(TestScope(), mockSession)
-    papaTracer.onWorkflowSessionStopped(123L)
+    tracer.onWorkflowSessionStarted(TestScope(), mockSession)
+    tracer.onWorkflowSessionStopped(123L)
   }
 
   @Test
@@ -221,13 +254,13 @@ internal class WorkflowPapaTracerTest {
 
     // Attach runtime context to tracer
     val testContext = TestRuntimeTraceContext()
-    papaTracer.attachRuntimeContext(testContext)
+    tracer.attachRuntimeContext(testContext)
 
     // Add session info to the context as would normally be done by WorkflowRuntimeMonitor
     testContext.workflowSessionInfo[mockSession.sessionId] = WorkflowSessionInfo(mockSession)
 
     // Should not throw
-    papaTracer.onRootPropsChanged(mockSession)
+    tracer.onRootPropsChanged(mockSession)
   }
 
   private class TestWorkflow : StatefulWorkflow<String, String, String, String>() {
