+++ workflow-tracing/api/workflow-tracing.api
@@ -130,7 +130,7 @@ public final class com/squareup/workflow1/tracing/RuntimeUpdates {
 	public final fun readAndClear ()Ljava/util/List;
 }
 
-public abstract interface class com/squareup/workflow1/tracing/SafeTraceInterface {
+public abstract interface class com/squareup/workflow1/tracing/SafeTraceInterface : com/squareup/workflow1/tracing/TraceInterface {
 	public abstract fun beginAsyncSection (Ljava/lang/String;I)V
 	public abstract fun beginSection (Ljava/lang/String;)V
 	public abstract fun endAsyncSection (Ljava/lang/String;I)V
@@ -146,6 +146,16 @@ public final class com/squareup/workflow1/tracing/SkipLogLine : com/squareup/wor
 	public fun hashCode ()I
 	public fun log (Ljava/lang/StringBuilder;)V
 	public fun toString ()Ljava/lang/String;
+}
+
+public abstract interface class com/squareup/workflow1/tracing/TraceInterface {
+	public abstract fun beginAsyncSection (Ljava/lang/String;I)V
+	public abstract fun beginSection (Ljava/lang/String;)V
+	public abstract fun endAsyncSection (Ljava/lang/String;I)V
+	public abstract fun endSection ()V
+	public abstract fun isCurrentlyTracing ()Z
+	public abstract fun isTraceable ()Z
+	public abstract fun logSection (Ljava/lang/String;)V
 }
 
 public final class com/squareup/workflow1/tracing/UiUpdateLogLine : com/squareup/workflow1/tracing/RuntimeUpdateLogLine {
